+++ packages/generate/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @safeql/generate
+# @ts-safeql/generate
 
 ## 0.0.1
 
@@ -6,5 +6,5 @@
 
 - initial release
 - Updated dependencies
-  - @safeql/shared@0.0.1
-  - @safeql/test-utils@0.0.1
+  - @ts-safeql/shared@0.0.1
+  - @ts-safeql/test-utils@0.0.1