+++ packages/test-utils/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @safeql/test-utils
+# @ts-safeql/test-utils
 
 ## 0.0.1
 
@@ -6,4 +6,4 @@
 
 - initial release
 - Updated dependencies
-  - @safeql/shared@0.0.1
+  - @ts-safeql/shared@0.0.1